Added "claim report" Backend functionality
bk_appReports was updated to correctly fetch reports from the database where the reports active tag is set to false.

The claim_report api endpoint was made. This will insert the bk_id and the r_id into the databases active_reports table. Then it will update the reports table to set the active tag true (removing it from the Home screen).